album and artist -> completed

+++ sorana/src/main/java/org/java/mentorship/controller/AlbumController.java
@@ -2,7 +2,10 @@ package org.java.mentorship.controller;
 
 import lombok.AllArgsConstructor;
 import org.java.mentorship.domain.Album;
+import org.java.mentorship.domain.Artist;
+import org.java.mentorship.domain.Song;
 import org.java.mentorship.service.AlbumService;
+import org.java.mentorship.service.SongService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
@@ -15,6 +18,7 @@ import static org.springframework.http.HttpStatus.CREATED;
 @AllArgsConstructor
 public class AlbumController {
     private final AlbumService albumService;
+    private final SongService songService;
 
     @GetMapping("/albums")
     public ResponseEntity<List<Album>> getAllalbums() {
@@ -22,8 +26,13 @@ public class AlbumController {
     }
 
     @GetMapping("/albums/{id}")
-    public ResponseEntity<Album> getArtistById(@PathVariable int id) {
+    public ResponseEntity<Album> getAlbumById(@PathVariable int id) {
         return ResponseEntity.ok(albumService.findById(id));
+    }
+
+    @GetMapping("/albums/{id}/songs")
+    public ResponseEntity<List<Song>> getSongsByAlbum(@PathVariable int id) {
+        return ResponseEntity.ok(songService.findSongsByAlbumId(id));
     }
 
     @PostMapping("/albums")
@@ -43,6 +52,10 @@ public class AlbumController {
 
     @DeleteMapping("/albums/{id}")
     public ResponseEntity<String> delete(@PathVariable("id") final int identifier) {
+        List<Song> songs = songService.findSongsByAlbumId(identifier);
+        if(!songs.isEmpty()) {
+            return ResponseEntity.badRequest().body("We need to delete all songs before deleting the album");
+        }
        albumService.delete(identifier);
         return ResponseEntity.ok("deleted album");
     }
